allow to have full control over exception handling
2 new protocols are added: `ExceptionHandler` and ExceptionDecoder

The former is used to wrap Exception handling, the later to
potentially decode exception into something else. So we have
decoupling on what/how.

All requests-* (and alike) now take an optional ExceptionHandler that
will dict how to deal with exceptions.

We provide a few impls of exception-handlers:

* `default-exception-handler` (the default): backward compatible,
throws on network + bad response codes
with vanilla exception types. basically 'identity'

* `decoding-exception-handler`: pass exceptions instance through
exceptions decoder and throw if we still have an exception instance
after decoding

* async-default-exception-handler and
  async-decoding-exception-handler, similar to the above, minus the
  throwing part since we're in callbacks/channels