Pass scroll parameters in a compatible way
Elasticsearch >= 2.0 changed the way pages are retrieved using scroll,
instead of just passing the scroll-id as body, it takes a JSON object.
This unfortunately fails when using older versions.

Thus it is also possible to continue passing the scroll-id as body and
passing the scrolling TTL as query parameter, which should allow both
newer and older versions of Elasticsearch to use the same request.

Fixes #12.