Add support for wrapping bulk chan payloads with metadata
This is a non-breaking change, which allows users to pass arbitrary data into the bulk-chan, and then make use of it on the output channel.
An example use case is correlating input with successes and failures. In our use case the elasticsearch ID may not be sufficient for this so extra metadata is required